pyproject.toml
[tool.poetry]
name = "novk"
version = "0.2.0"
description = ""
authors = ["Eduard Gan <eduard.gan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.2"


django = {version = "4.2rc1", allow-prereleases = true}
django-environ = "0.10.0"
django-widget-tweaks = "1.4.12"
django_debug_toolbar = "3.8.1"

mutagen = "1.46.0"

psycopg2-binary = "2.9.5"

gunicorn = "20.1.0"
uvicorn = "0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
